Move and rename data-site-map
